media: Pass struct device pointer to fwnode endpoint parsing functions
The v4l2_fwnode_endpoint_parse() and v4l2_fwnode_endpoint_alloc_parse()
functions print lots of debugging information with pr_debug(), which can
make the log confusing as the information isn't associated with a
device. Pass a struct device pointer to the functions to prepare for
replacing the pr_debug() calls with dev_dbg(). No functional change is
intended in this commit.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>