media: i2c: imx219: Add internal sink pads
Augment the imx219 media entity with 2 internal sink pads: one for
image data, the second for embedded data. This prepares to add support
for routing and stream multiplexing.

To maintain compatibility with existing applications assign index 1 and
2 to the newly added pads, maintaining the existing source pad at index 0.

For the same reason, disable setting format on the internal image sink
pad and propagate the format -backward- from the source pad to the image
sink pad. This contradict the APIs but guarantee existing applications
still work as they used to.

Signed-off-by: Jacopo Mondi <jacopo.mondi@ideasonboard.com>